feat: MVP of the source functionality.
StaticFileSource finds the file indicated by target of the request in the root folder of the provided ServerConfig.
The instance of the StaticFileSource keeps data on the read file (size, read data in vector<char>, MIME type/subtype, HTTP code).
The instance can be then used to for HTTP response.
Basic tests passed.

Backlog:
-search for matching Location;
-setting MIME type/subtype;
-generating index page;
-checking the correct return of Error page;
-checking correct redirection;